back/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose[cryptography]
bcrypt
pydantic
pydantic[email]
requests
pymongo